+++ src/includes/exec.h
@@ -1,15 +1,14 @@
 #ifndef EXEC_H
-#define EXEC_H
+# define EXEC_H
 
-#include "ast.h"
-#include "type.h"
+# include "ast.h"
+# include "type.h"
 
 typedef struct s_pids
 {
 	pid_t			pid;
 	struct s_pids	*next;
 }					t_pids;
-
 
 void				exec_heredoc(t_astnode *node);
 pid_t				fork_and_exec_child(t_astnode *node, int old_pipes[2],
@@ -25,6 +24,8 @@ pid_t				*create_pids(int n);
 int					**create_pipes(int n);
 int					is_directory(char *str);
 int					is_command(char *str);
-
+t_status			execute_one_builtin(t_astnode *root);
+t_status			execute_fork_commands(t_list *cmds);
+void				waitpids(t_list *cmds, pid_t *pids);
 
 #endif
