+++ src/utils/grobal_vars1.c
@@ -40,50 +40,6 @@ t_env	*grobal_envlist(int get_or_set, ...)
 	return (g_env);
 }
 
-void	update_grobal_envlist(void)
-{
-	t_env	*env;
-
-	env = grobal_envlist(GET);
-	free_env_list(env);
-	env = create_env_list(grobal_env(GET));
-	grobal_envlist(SET, env);
-	return ;
-}
-
-char	**create_env(t_env *envlist)
-{
-	char	**env;
-	int		len;
-	int		i;
-	t_env	*node;
-	char	*value;
-
-	len = env_size(envlist);
-	env = (char **)malloc(sizeof(char *) * (len + 1));
-	node = envlist;
-	i = 0;
-	while (i < len)
-	{
-		value = node->value;
-		env[i++] = ft_strdup(value);
-		node = node->next;
-	}
-	env[i] = NULL;
-	return (env);
-}
-
-void	update_grobal_env(void)
-{
-	char	**env;
-
-	env = grobal_env(GET);
-	ft_2darraydel(env);
-	env = create_env(grobal_envlist(GET));
-	grobal_env(SET, env);
-	return ;
-}
-
 char	*grobal_tmpfile(int get_or_set, ...)
 {
 	va_list		ap;
